+++ src/main/java/dev/willyelton/crystal_tools/config/CrystalToolsConfig.java
@@ -41,6 +41,7 @@ public class CrystalToolsConfig {
     public static ForgeConfigSpec.BooleanValue ENABLE_ITEM_REQUIREMENTS;
     public static ForgeConfigSpec.DoubleValue BACKGROUND_OPACITY;
     public static ForgeConfigSpec.BooleanValue PAUSE_SCREEN;
+    public static ForgeConfigSpec.BooleanValue REQUIRE_CRYSTAL_FOR_RESET;
 
     static {
         ForgeConfigSpec.Builder configBuilder = new ForgeConfigSpec.Builder();
@@ -122,5 +123,8 @@ public class CrystalToolsConfig {
 
         PAUSE_SCREEN = builder.comment("If it is true then the skill tree screen pauses the game (in single-player), if false then it doesn't")
                 .define("pause_screen", true);
+
+        REQUIRE_CRYSTAL_FOR_RESET = builder.comment("Require a crystal item in your inventory for resetting skill points")
+                .define("require_crystal_for_reset", true);
     }
 }